+++ src/main/groovy/nebula/plugin/publishing/maven/MavenResolvedDependenciesPlugin.groovy
@@ -17,6 +17,10 @@ package nebula.plugin.publishing.maven
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.XmlProvider
+import org.gradle.api.artifacts.component.ModuleComponentSelector
+import org.gradle.api.artifacts.result.ResolvedDependencyResult
+import org.gradle.api.publish.maven.MavenPublication
 
 /**
  * Replaces dynamic constraints with fixed constraints
@@ -26,6 +30,43 @@ class MavenResolvedDependenciesPlugin implements Plugin<Project> {
     void apply(Project project) {
         project.plugins.apply(MavenDependenciesPlugin)
 
-        new MavenResolvedDependenciesBase().addResolveRules(project)
+        project.publishing {
+            publications {
+                nebula(MavenPublication) {
+                    pom.withXml { XmlProvider xml->
+                        def dependencies = xml.asNode()?.dependencies?.dependency
+                        def dependencyMap = [:]
+
+                        dependencyMap['runtime'] = project.configurations.runtime.incoming.resolutionResult.allDependencies
+                        dependencyMap['test'] = project.configurations.testRuntime.incoming.resolutionResult.allDependencies - dependencyMap['runtime']
+                        dependencies?.each { Node dep ->
+                            def group = dep.groupId.text()
+                            def name = dep.artifactId.text()
+                            def scope = dep.scope.text()
+
+                            if (scope == 'provided') {
+                                scope = 'runtime'
+                            }
+
+                            ResolvedDependencyResult resolved = dependencyMap[scope].find { r ->
+                                (r.requested instanceof ModuleComponentSelector) &&
+                                        (r.requested.group == group) &&
+                                        (r.requested.module == name)
+                            }
+
+                            if (!resolved) {
+                                return  // continue loop if a dependency is not found in dependencyMap
+                            }
+
+                            def versionNode = dep.version
+                            if (!versionNode) {
+                                versionNode = dep.appendNode('version')
+                            }
+                            dep.version[0].value = resolved?.selected?.moduleVersion?.version
+                        }
+                    }
+                }
+            }
+        }
     }
 }